enrichment/src/requirements.txt
functions-framework==3.*
firebase-functions==0.4.2
jsonschema==4.19.0
google-auth-oauthlib==0.8.0
google-cloud-storage==3.0.0
google-cloud-bigquery==3.29.0
google-genai==1.5.0
pandas==2.2.3
fsspec==2025.2.0
gcsfs==2025.2.0